Move event_retriever to grapl common
