+++ src/python/grapl-common/grapl_common/sqs/event_retriever.py
